requirements.txt
# Fichier requirements.txt
python-dotenv == 1.0.1
tqdm==4.66.2
ultralytics==8.1.34
#pip install -r requirements.txt
#numpy.__version__ || pip list || pip show numpy || pip freeze | findstr numpy